SRSENB: Added PUSCH TA and EVM measurement. Some more PHY cleanup.

+++ lib/include/srslte/phy/channel/delay.h
@@ -32,7 +32,7 @@ typedef struct {
   uint32_t srate_max_hz;
   uint32_t srate_hz;
   float    delay_us;
-  float    delay_nsamples;
+  uint32_t delay_nsamples;
 
   srslte_ringbuffer_t rb;
   cf_t*               zero_buffer;
